Stat of "fetcherfying"
Endpoint clean up
/metadata/document_id/{id} -> /metadata/document/{id}

/authorship update by admin has audits. paper password update has an admin audit.

I apparently made a broken change in submission listing. "status" used to be incorrectly implemented as dual purpose - which was working. I broke it, and became a ticket.  Submission status moniker is a new query param. submission_status_group - which is what it is. No more hitching to generic "status".